requirements.txt
azure-functions
requests
pytz
schedule